backend/requirements.txt
fastapi
uvicorn
fastf1
pandas
google-generativeai
python-dotenv
